COMN-65: fix JSON processing of list extended attributes.
The theory of extended properties is that, in JSON, they
look like all other properties. This is what makes the
compatibility work right. If, today, "aaa" is an extended
prop, and tomorrow it becomes part of the data model,
we want json from today to work. So, we want the ext
props in json to be ordinary, top-level props, not
pushed down into another level of json object. For example:

```
{
  "data" : "Cuthbert Girdlestone",
  "attributes" : {
    "entityMentions" : {
      "type" : "list",
      "itemType" : "entityMentions",
      "items" : [ {
        "startOffset" : 0,
        "endOffset" : 20,
        "entityType" : "PERSON",
        "coreferenceChainId" : -1,
        "confidence" : 0.0,
        "flags" : 0
      } ],
      "extra_key" : "extra_value"
    }
  },
  "documentMetadata" : { }
}
```